Message Widgets for Pickups
Created Message Widgets that appear when the player pickups up a potion or crystal.

When an effect is applied by picking up a potion, we broadcast the tags on that potion. The overlay listens for that broadcast and finds a message in our data table that matches the tag. We then broadcast the matching table row which allows us to use that data in blueprint to generate a UI message widget that appears on screen with the appropriate icon and message.